Rename Sort -> SortScratch
